android_activity

Module input

Structs

ButtonState

A bitfield representing the state of buttons during a motion event.

EdgeFlags

A bitfield representing which edges were touched by a motion event.

InputIterator

An exclusive, lending iterator for input events

KeyCharacterMap

Describes the keys provided by a keyboard device and their associated labels.

KeyEvent

A key event

KeyEventFlags

Flags associated with KeyEvent.

MetaState

A bitfield representing the state of modifier keys during an event.

MotionEvent

A motion event

MotionEventFlags

Flags associated with this MotionEvent.

Pointer

A view into the data of a specific pointer in a motion event.

PointersIter

An iterator over the pointers in a MotionEvent.

TextInputState

TextSpan

This struct holds a span within a region of text from start to end.

Enums

Axis

An axis of a motion event.

Button

Identifies buttons that are associated with motion events.

InputEvent

Enum of possible input events

KeyAction

Key actions.

KeyMapChar

Either represents, a unicode character or combining accent from a KeyCharacterMap, or None for non-printable keys.

KeyboardType

An enum representing the types of keyboards that may generate key events

Keycode

Key codes.

MotionAction

A motion action.

Source

An enum representing the source of an MotionEvent or KeyEvent

ToolType

The tool type of a pointer.