android_activity

Struct ConfigurationRef

pub struct ConfigurationRef { /* private fields */ }

A (cheaply clonable) reference to this applicationโs ndk::configuration::Configuration

This provides a thread-safe way to access the latest configuration state for an application without deeply copying the large ndk::configuration::Configuration struct.

If the application is notified of configuration changes then those changes will become visible via pre-existing configuration references.

Implementations

impl ConfigurationRef

pub fn copy(&self) -> Configuration

pub fn country(&self) -> Option<String>

Returns the country code, as a String of two characters, if set

pub fn density(&self) -> Option<u32>

Returns the screen density in dpi.

On some devices it can return values outside of the density enum.

pub fn keyboard(&self) -> Keyboard

Returns the keyboard type.

pub fn keys_hidden(&self) -> KeysHidden

Returns keyboard visibility/availability.

pub fn language(&self) -> Option<String>

Returns the language, as a String of two characters, if a language is set

pub fn layout_direction(&self) -> LayoutDir

Returns the layout direction

pub fn mcc(&self) -> i32

Returns the mobile country code.

pub fn mnc(&self) -> Option<i32>

Returns the mobile network code, if one is defined

pub fn nav_hidden(&self) -> NavHidden

pub fn navigation(&self) -> Navigation

pub fn orientation(&self) -> Orientation

pub fn screen_height_dp(&self) -> Option<i32>

pub fn screen_width_dp(&self) -> Option<i32>

pub fn screen_long(&self) -> ScreenLong

pub fn screen_size(&self) -> ScreenSize

pub fn sdk_version(&self) -> i32

pub fn smallest_screen_width_dp(&self) -> Option<i32>

pub fn touchscreen(&self) -> Touchscreen

pub fn ui_mode_night(&self) -> UiModeNight

pub fn ui_mode_type(&self) -> UiModeType